MM-93 Simulate 1 second Calls
Commits were lost due to deleting a branch which was deleted to fix a Rebasing issue. Saved the code before deleting so now all that is changed is all at once.
Added news variables that load from server call to client side to use as a point to track where user should be in a second
Add function render progress which renders progress and bar width given progress and duration
Added to showNothingPLaying to handle when user is not playing a song
Added tick function to call every second to simulate a real call to spotify